Add loading screen during db import
Rather than loading the db on the main thread and stalling out hard,
let's throw up a loading screen while we load the db. Updates via an
event handler to show that we're making progress.

Next steps: progress bar for the Sauvegarde import, since that's the
longest step.

TODO: maybe not use Lazy KTUserContext _Db as the mechanism here, it's
brittle if execution order changes. Switch to an explicit Initialize
step, before which KTContext.Db is null.